refactor(library/math): moves `cofactor`
