+++ src/library/math/operator/constructive/cofactor.ts
@@ -1,6 +1,18 @@
 import {
-  cofactor,
-} from '../../../math/operators';
+  leastCommonMultiple,
+} from './leastCommonMultiple';
+
+const cofactor = (
+  {
+    to: givenCofactor,
+    forLCMWith: givenValue,
+  }: {
+    to: number;
+    forLCMWith: number;
+  },
+): number => {
+  return leastCommonMultiple(givenValue, givenCofactor) / givenCofactor;
+};
 
 export {
   cofactor,
